+++ reminder/reminder/settings/common.py
@@ -50,7 +50,7 @@ THIRD_PARTY_APPS = (
 )
 
 LOCAL_APPS = (
-    'tasks',
+    'event_reminder_app',
 )
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
