+++ reminder/reminder/urls.py
@@ -18,5 +18,5 @@ from django.contrib import admin
 
 urlpatterns = [
     url(r'^admin/', include(admin.site.urls)),
-    url(r'', include("tasks.urls")),
+    url(r'', include("event_reminder_app.urls")),
 ]
